tools/gcc-plugins: import rename
Move files into proper place for use in FRR.
